Time: 27 ms (36.48%), Space: 46.6 MB (6.76%) - LeetHub

+++ 0518-coin-change-ii/0518-coin-change-ii.cpp
@@ -21,6 +21,6 @@ public:
             }
         }
 
-        return static_cast<int>(dp[n][m]);
+        return dp[n][m];
     }
 };
